Move CreditCardPayButton to own file
